+++ src/services/ManellicTree.jl
@@ -46,30 +46,6 @@ function eigenCoords(
 end
 
 
-# test 
-function testEigenCoords(
-  r_C = pi/3,
-  ax_CC = [SA[5*randn();randn()] for _ in 1:100],
-)
-  M = TranslationGroup(2)
-  _R(α, s=exp(-α*im)) = real(s)*SA[1 0; 0 1] + imag(s)*SA[0 1; -1 0]
-  # _R(α) = SA[cos(α) sin(α); -sin(α) cos(α)]
-  r_R_ax = _R(r_C)
-  # rotate coordinates
-  r_CC = map(ax_CC) do ax_C
-    r_R_ax*ax_C + SA[10;-100]
-  end
-  r_CV = Manifolds.cov(M, r_CC)
-  r_R_ax_, L, pidx = eigenCoords(r_CV)
-
-  # spot check
-  _ax_ERR = log_lie(SpecialOrthogonal(2), (r_R_ax_')*r_R_ax)[1,2]
-  @assert isapprox(_ax_ERR; atol = 5/length(ax_CC)) "Spot check failed on eigen split of manifold points, the estimated point rotation matrix did not match construction."
-
-  r_CC, r_R_ax_, pidx, r_CV
-end
-
-
 
 # give vector of manifold points and split along largest covariance (i.e. major direction)
 function splitPointsEigen(
